Add eslint, and break up the behemoth file into smaller files
Summary:
This'll be much easier to understand and maintain, I hope.

Test plan:
`yarn flow` and `yarn eslint .` should be happy.
