+++ .eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+    extends: ['@khanacademy'],
+    root: true,
+    parser: '@babel/eslint-parser',
+    rules: {
+        'prettier/prettier': ['error', {singleQuote: true}],
+    },
+};
