+++ package.json
@@ -4,6 +4,14 @@
         "test": "jest"
     },
     "devDependencies": {
+        "@babel/eslint-parser": "^7.17.0",
+        "@khanacademy/eslint-config": "^0.1.0",
+        "eslint": "8.7.0",
+        "eslint-config-prettier": "7.0.0",
+        "eslint-plugin-jsx-a11y": "^6.5.1",
+        "eslint-plugin-prettier": "^4.0.0",
+        "eslint-plugin-react": "^7.29.2",
+        "eslint-plugin-react-hooks": "^4.3.0",
         "flow-bin": "^0.172.0",
         "jest": "^27.5.1"
     },
@@ -16,8 +24,10 @@
         "@babel/types": "^7.17.0",
         "apollo-utilities": "^1.3.4",
         "babel-jest": "23.4.2",
+        "eslint-plugin-flowtype": "^8.0.3",
         "graphql": "14.2.1",
         "graphql-tag": "2.10.1",
+        "prettier": "^2.5.1",
         "prettier-eslint": "^13.0.0"
     }
 }
